pyproject.toml
[tool.poetry]
name = "events_publisher"
version = "0.1.0"
description = ""
authors = ["Pedro Rivera <pedro.rivera@globant.com>"]

[tool.poetry.dependencies]
python = "^3.9"
kafka-python = "^2.0.2"

[tool.poetry.dev-dependencies]
pytest = "^5.2"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
